// src/components/FormComponents/types.ts
export type TMutiTextDefaultType = string[];

export type TMutiTextObjectType = {
  text: string;
}[];